+++ examples/catalogs/declarative_injections.py
@@ -27,16 +27,13 @@ class Services(catalogs.DeclarativeCatalog):
     """Catalog of service providers."""
 
     database = providers.Singleton(sqlite3.connect, ':memory:')
-    """:type: providers.Provider -> sqlite3.Connection"""
 
     users = providers.Factory(UsersService,
                               db=database)
-    """:type: providers.Provider -> UsersService"""
 
     auth = providers.Factory(AuthService,
                              db=database,
                              users_service=users)
-    """:type: providers.Provider -> AuthService"""
 
 
 # Retrieving service providers from catalog: