+++ examples/catalogs/override_declarative_by_dynamic.py
@@ -18,11 +18,9 @@ class Catalog(catalogs.DeclarativeCatalog):
     object1_factory = providers.Factory(Object1,
                                         arg1=1,
                                         arg2=2)
-    """:type: providers.Provider -> Object1"""
 
     object2_factory = providers.Factory(Object2,
                                         object1=object1_factory)
-    """:type: providers.Provider -> Object2"""
 
 
 # Overriding `Catalog` with some `DynamicCatalog` instance: