+++ examples/miniapps/api_client/models.py
@@ -1,0 +1,14 @@
+"""asd."""
+
+
+class User(object):
+    """User model."""
+
+    def __init__(self, id, api_client):
+        """Initializer."""
+        self.id = id
+        self.api_client = api_client
+
+    def register(self):
+        """Register user."""
+        self.api_client.call('register', {'id': self.id})