+++ examples/miniapps/api_client/tests.py
@@ -1,0 +1,12 @@
+"""asd."""
+
+from mock import Mock
+
+import main
+import api
+
+# Mock ApiClient for testing:
+with main.api_client.override(Mock(api.ApiClient)) as api_client_mock:
+    user = main.user_factory('test')
+    user.register()
+    api_client_mock().call.assert_called_with('register', {'id': 'test'})