+++ examples/miniapps/flask_services/app.py
@@ -1,0 +1,10 @@
+"""Flask application."""
+
+import flask
+
+
+app = flask.Flask(__name__)
+
+
+if __name__ == '__main__':
+    app.run()