+++ examples/miniapps/flask_services/services/__init__.py
@@ -1,0 +1,8 @@
+"""Services package."""
+
+from dependency_injector import catalogs
+from dependency_injector import providers
+
+
+class ServicesModule(catalogs.DeclarativeCatalog):
+    """Service providers module."""