+++ docs/main/changelog.rst
@@ -9,6 +9,7 @@ follows `Semantic versioning`_
 
 Development version
 -------------------
+- No features.
 
 1.15.2
 ------